# chunk/codec: adjust decoding offsets in chunk